Moved class "Brush" into Brush.js
